requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pyquery
curl_cffi
bencoder
loguru
apscheduler
requests
jinja2
python-jose
passlib
qbittorrent-api
transmission_rpc
protobuf
grpcio